refactor: relocate utils
