+++ integrations/GA4/browser.js
@@ -10,7 +10,7 @@ import {
   getPageViewProperty,
   hasRequiredParameters,
 } from "./utils";
-import { type, flattenJsonPayload } from "../../utils/utils";
+import { type, flattenJsonPayload } from "../utils/commonUtils";
 import { NAME } from "./constants";
 
 export default class GA4 {
