+++ integrations/Klaviyo/browser.js
@@ -3,7 +3,7 @@
 import get from "get-value";
 import logger from "../../utils/logUtil";
 import ScriptLoader from "../ScriptLoader";
-import { extractCustomFields, getDefinedTraits } from "../../utils/utils";
+import { extractCustomFields, getDefinedTraits } from "../utils/commonUtils";
 import ecommEventPayload from "./util";
 import { isNotEmpty } from "../utils/commonUtils";
 import { NAME } from "./constants";
