+++ integrations/Sentry/browser.js
@@ -7,7 +7,7 @@ import get from "get-value";
 import logger from "../../utils/logUtil";
 import { SentryScriptLoader, sentryInit } from "./utils";
 import { removeUndefinedAndNullValues } from "../utils/commonUtils";
-import { getDefinedTraits, isObject } from "../../utils/utils";
+import { getDefinedTraits, isObject } from "../utils/commonUtils";
 import { NAME } from "./constants";
 
 class Sentry {
