+++ utils/utils.js
@@ -1,4 +1,4 @@
-// import * as XMLHttpRequestNode from "Xmlhttprequest";
+/* eslint-disable no-bitwise */
 import { parse } from "component-url";
 import get from "get-value";
 import { v4 as uuid } from "@lukeed/uuid";
@@ -9,6 +9,54 @@ import { clientToServerNames } from "./client_server_name";
 import { CONFIG_URL, RESERVED_KEYS } from "./constants";
 import Storage from "./storage";
 
+
+/**
+ * This function is to send handled errors to Bugsnag if Bugsnag client is available
+ * @param {Error} error Error instance from handled error
+ */
+ function notifyError(error) {
+  if (window.rsBugsnagClient) {
+    window.rsBugsnagClient.notify(error);
+  }
+}
+
+function handleError(error, analyticsInstance) {
+  let errorMessage = error.message;
+  try {
+    if (error instanceof Event) {
+      // Discard all the non-script loading errors
+      if (error.target && error.target.localName !== "script") return;
+
+      // Discard errors of scripts that are not loaded by the SDK
+      if (error.target.dataset && error.target.dataset.loader !== LOAD_ORIGIN)
+        return;
+
+      errorMessage = `error in script loading:: src::  ${error.target.src} id:: ${error.target.id}`;
+
+      // SDK triggered ad-blocker script
+      if (error.target.id === "ad-block") {
+        analyticsInstance.page(
+          "RudderJS-Initiated",
+          "ad-block page request",
+          { path: "/ad-blocked", title: errorMessage },
+          analyticsInstance.sendAdblockPageOptions
+        );
+        // No need to proceed further for Ad-block errors
+        return;
+      }
+    }
+
+    errorMessage = `[handleError]:: "${errorMessage}"`;
+    logger.error(errorMessage);
+    let errorObj = error;
+    if (!(error instanceof Error)) errorObj = new Error(errorMessage);
+    notifyError(errorObj);
+  } catch (err) {
+    logger.error("[handleError] Exception:: ", err);
+    notifyError(err);
+  }
+}
+
 /**
  *
  * Utility method for excluding null and empty values in JSON
@@ -18,7 +66,7 @@ import Storage from "./storage";
  */
 function replacer(key, value) {
   if (value === null || value === undefined) {
-    return undefined;
+    return;
   }
   return value;
 }
@@ -72,9 +120,9 @@ function getJSON(url, wrappers, isLoaded, callback) {
   const xhr = new XMLHttpRequest();
 
   xhr.open("GET", url, false);
-  xhr.onload = function () {
+  xhr.onload = () => {
     const { status } = xhr;
-    if (status == 200) {
+    if (status === 200) {
       // logger.debug("status 200");
       callback(null, xhr.responseText, wrappers, isLoaded);
     } else {
@@ -93,7 +141,7 @@ function getJSON(url, wrappers, isLoaded, callback) {
  */
 function getJSONTrimmed(context, url, writeKey, callback) {
   // server-side integration, XHR is node module
-  const cb_ = callback.bind(context);
+  const cb = callback.bind(context);
 
   const xhr = new XMLHttpRequest();
 
@@ -104,19 +152,42 @@ function getJSONTrimmed(context, url, writeKey, callback) {
     // `Basic ${Buffer.from(`${writeKey}:`).toString("base64")}`
   );
 
-  xhr.onload = function () {
+  xhr.onload = () => {
     const { status } = xhr;
-    if (status == 200) {
+    if (status === 200) {
       // logger.debug("status 200 " + "calling callback");
-      cb_(200, xhr.responseText);
+      cb(200, xhr.responseText);
     } else {
       handleError(
         new Error(`request failed with status: ${xhr.status} for url: ${url}`)
       );
-      cb_(status);
+      cb(status);
     }
   };
   xhr.send();
+}
+
+function getReferrer() {
+  return document.referrer || "$direct";
+}
+
+function getReferringDomain(referrer) {
+  const split = referrer.split("/");
+  if (split.length >= 3) {
+    return split[2];
+  }
+  return "";
+}
+
+function getCanonicalUrl() {
+  const tags = document.getElementsByTagName("link");
+  for (let i = 0; i < tags.length; i += 1) {
+    const tag = tags[i];
+    if (!tag) break;
+    if (tag.getAttribute("rel") === "canonical") {
+      return tag.getAttribute("href");
+    }
+  }
 }
 
 /**
@@ -129,51 +200,14 @@ function leaveBreadcrumb(breadcrumb) {
   }
 }
 
-/**
- * This function is to send handled errors to Bugsnag if Bugsnag client is available
- * @param {Error} error Error instance from handled error
- */
-function notifyError(error) {
-  if (window.rsBugsnagClient) {
-    window.rsBugsnagClient.notify(error);
+function getUrl(search) {
+  const canonicalUrl = getCanonicalUrl();
+  let url = window.location.href;
+  if (canonicalUrl) {
+    url = canonicalUrl.indexOf("?") > -1 ? canonicalUrl : canonicalUrl + search;
   }
-}
-
-function handleError(error, analyticsInstance) {
-  let errorMessage = error.message;
-  try {
-    if (error instanceof Event) {
-      // Discard all the non-script loading errors
-      if (error.target && error.target.localName !== "script") return;
-
-      // Discard errors of scripts that are not loaded by the SDK
-      if (error.target.dataset && error.target.dataset.loader !== LOAD_ORIGIN)
-        return;
-
-      errorMessage = `error in script loading:: src::  ${error.target.src} id:: ${error.target.id}`;
-
-      // SDK triggered ad-blocker script
-      if (error.target.id === "ad-block") {
-        analyticsInstance.page(
-          "RudderJS-Initiated",
-          "ad-block page request",
-          { path: "/ad-blocked", title: errorMessage },
-          analyticsInstance.sendAdblockPageOptions
-        );
-        // No need to proceed further for Ad-block errors
-        return;
-      }
-    }
-
-    errorMessage = `[handleError]:: "${errorMessage}"`;
-    logger.error(errorMessage);
-    let errorObj = error;
-    if (!(error instanceof Error)) errorObj = new Error(errorMessage);
-    notifyError(errorObj);
-  } catch (err) {
-    logger.error("[handleError] Exception:: ", err);
-    notifyError(err);
-  }
+  const hashIndex = url.indexOf("#");
+  return hashIndex > -1 ? url.slice(0, hashIndex) : url;
 }
 
 function getDefaultPageProperties() {
@@ -185,103 +219,45 @@ function getDefaultPageProperties() {
   const { search } = window.location;
   const { title } = document;
   const url = getUrl(search);
+  // eslint-disable-next-line camelcase
   const tab_url = window.location.href;
 
   const referrer = getReferrer();
+  // eslint-disable-next-line camelcase
   const referring_domain = getReferringDomain(referrer);
+  // eslint-disable-next-line camelcase
   const initial_referrer = Storage.getInitialReferrer();
+  // eslint-disable-next-line camelcase
   const initial_referring_domain = Storage.getInitialReferringDomain();
   return {
     path,
     referrer,
+    // eslint-disable-next-line camelcase
     referring_domain,
     search,
     title,
     url,
+    // eslint-disable-next-line camelcase
     tab_url,
+    // eslint-disable-next-line camelcase
     initial_referrer,
+    // eslint-disable-next-line camelcase
     initial_referring_domain,
   };
 }
 
-function getReferrer() {
-  // This error handling is in place to avoid accessing dead object(document)
-  const defaultReferrer = "$direct";
-  try {
-    return document.referrer || defaultReferrer;
-  } catch (e) {
-    logger.error("Error trying to access 'document.referrer': ", e);
-    return defaultReferrer;
-  }
-}
-
-function getReferringDomain(referrer) {
-  const split = referrer.split("/");
-  if (split.length >= 3) {
-    return split[2];
-  }
-  return "";
-}
-
-function getUrl(search) {
-  const canonicalUrl = getCanonicalUrl();
-  const url = canonicalUrl
-    ? canonicalUrl.indexOf("?") > -1
-      ? canonicalUrl
-      : canonicalUrl + search
-    : window.location.href;
-  const hashIndex = url.indexOf("#");
-  return hashIndex > -1 ? url.slice(0, hashIndex) : url;
-}
-
-function getCanonicalUrl() {
-  const tags = document.getElementsByTagName("link");
-  for (var i = 0, tag; (tag = tags[i]); i++) {
-    if (tag.getAttribute("rel") === "canonical") {
-      return tag.getAttribute("href");
-    }
-  }
-}
-
-function getCurrency(val) {
-  if (!val) return;
-  if (typeof val === "number") {
-    return val;
-  }
-  if (typeof val !== "string") {
-    return;
-  }
-
-  val = val.replace(/\$/g, "");
-  val = parseFloat(val);
-
-  if (!isNaN(val)) {
-    return val;
-  }
-}
-
-function getRevenue(properties, eventName) {
-  let { revenue } = properties;
-  const orderCompletedRegExp =
-    /^[ _]?completed[ _]?order[ _]?|^[ _]?order[ _]?completed[ _]?$/i;
-
-  // it's always revenue, unless it's called during an order completion.
-  if (!revenue && eventName && eventName.match(orderCompletedRegExp)) {
-    revenue = properties.total;
-  }
-
-  return getCurrency(revenue);
-}
-
 function transformNamesCore(integrationObject, namesObj) {
   Object.keys(integrationObject).forEach((key) => {
-    if (integrationObject.hasOwnProperty(key)) {
+    if (integrationObject[key]) {
       if (namesObj[key]) {
+        // eslint-disable-next-line no-param-reassign
         integrationObject[namesObj[key]] = integrationObject[key];
       }
-      if (key != "All") {
-        // delete user supplied keys except All and if except those where oldkeys are not present or oldkeys are same as transformed keys
-        if (namesObj[key] != undefined && namesObj[key] != key) {
+      if (key !== "All") {
+        // delete user supplied keys except All and if except those where
+        // old keys are not present or old keys are same as transformed keys
+        if (namesObj[key] !== undefined && namesObj[key] !== key) {
+          // eslint-disable-next-line no-param-reassign
           delete integrationObject[key];
         }
       }
@@ -367,78 +343,6 @@ function findAllEnabledDestinations(
   return enabledList;
 }
 
-/**
- * reject all null values from array/object
- * @param  {} obj
- * @param  {} fn
- */
-function rejectArr(obj, fn) {
-  fn = fn || compact;
-  return type(obj) == "array" ? rejectarray(obj, fn) : rejectobject(obj, fn);
-}
-
-/**
- * particular case when rejecting an array
- * @param  {} arr
- * @param  {} fn
- */
-var rejectarray = function (arr, fn) {
-  const ret = [];
-
-  for (let i = 0; i < arr.length; ++i) {
-    if (!fn(arr[i], i)) ret[ret.length] = arr[i];
-  }
-
-  return ret;
-};
-
-/**
- * Rejecting null from any object other than arrays
- * @param  {} obj
- * @param  {} fn
- *
- */
-var rejectobject = function (obj, fn) {
-  const ret = {};
-
-  for (const k in obj) {
-    if (obj.hasOwnProperty(k) && !fn(obj[k], k)) {
-      ret[k] = obj[k];
-    }
-  }
-
-  return ret;
-};
-
-function compact(value) {
-  return value == null;
-}
-
-/**
- * check type of object incoming in the rejectArr function
- * @param  {} val
- */
-function type(val) {
-  switch (Object.prototype.toString.call(val)) {
-    case "[object Function]":
-      return "function";
-    case "[object Date]":
-      return "date";
-    case "[object RegExp]":
-      return "regexp";
-    case "[object Arguments]":
-      return "arguments";
-    case "[object Array]":
-      return "array";
-  }
-
-  if (val === null) return "null";
-  if (val === undefined) return "undefined";
-  if (val === Object(val)) return "object";
-
-  return typeof val;
-}
-
 function getUserProvidedConfigUrl(configUrl, defConfigUrl) {
   let url = configUrl;
   if (url.indexOf("sourceConfig") === -1) {
@@ -454,219 +358,34 @@ function getUserProvidedConfigUrl(configUrl, defConfigUrl) {
   }
   return url;
 }
+
+/**
+ * Check if a reserved keyword is present in the given object
+ * @param {*} inpObj
+ * @param {*} msgType
+ */
+function checkForReservedKeywords(inpObj, msgType) {
+  if (inpObj) {
+    Object.keys(inpObj).forEach((key) => {
+      if (RESERVED_KEYS.includes(key.toLowerCase())) {
+        logger.error(`Reserved keyword '${key}' is used in '${msgType}' call`);
+      }
+    });
+  }
+}
+
 /**
  * Check if a reserved keyword is present in properties/traits
- * @param {*} properties
- * @param {*} reservedKeywords
- * @param {*} type
- */
-function checkReservedKeywords(message, messageType) {
-  //  properties, traits, contextualTraits are either undefined or object
-  const { properties, traits } = message;
-  if (properties) {
-    Object.keys(properties).forEach((property) => {
-      if (RESERVED_KEYS.indexOf(property.toLowerCase()) >= 0) {
-        logger.error(
-          `Warning! : Reserved keyword used in properties--> ${property} with ${messageType} call`
-        );
-      }
-    });
-  }
-  if (traits) {
-    Object.keys(traits).forEach((trait) => {
-      if (RESERVED_KEYS.indexOf(trait.toLowerCase()) >= 0) {
-        logger.error(
-          `Warning! : Reserved keyword used in traits--> ${trait} with ${messageType} call`
-        );
-      }
-    });
-  }
-  const contextualTraits = message.context.traits;
-  if (contextualTraits) {
-    Object.keys(contextualTraits).forEach((contextTrait) => {
-      if (RESERVED_KEYS.indexOf(contextTrait.toLowerCase()) >= 0) {
-        logger.error(
-          `Warning! : Reserved keyword used in traits --> ${contextTrait} with ${messageType} call`
-        );
-      }
-    });
-  }
-}
-
-/* ------- Start FlattenJson -----------
- * This function flatten given json object to single level.
- * So if there is nested object or array, all will appear in first level properties of an object.
- * Following is case we are handling in this function ::
- * condition 1: String
- * condition 2: Array
- * condition 3: Nested object
- */
-function recurse(cur, prop, result) {
-  const res = result;
-  if (Object(cur) !== cur) {
-    res[prop] = cur;
-  } else if (Array.isArray(cur)) {
-    const l = cur.length;
-    for (let i = 0; i < l; i += 1)
-      recurse(cur[i], prop ? `${prop}.${i}` : `${i}`, res);
-    if (l === 0) res[prop] = [];
-  } else {
-    let isEmpty = true;
-    Object.keys(cur).forEach((key) => {
-      isEmpty = false;
-      recurse(cur[key], prop ? `${prop}.${key}` : key, res);
-    });
-    if (isEmpty) res[prop] = {};
-  }
-  return res;
-}
-
-function flattenJsonPayload(data) {
-  return recurse(data, "", {});
-}
-/* ------- End FlattenJson ----------- */
-/**
- *
  * @param {*} message
- * @param {*} destination
- * @param {*} keys
- * @param {*} exclusionFields
- * Extract fields from message with exclusions
- * Pass the keys of message for extraction and
- * exclusion fields to exclude and the payload to map into
- * -----------------Example-------------------
- * extractCustomFields(message,payload,["traits", "context.traits", "properties"], "email",
- * ["firstName",
- * "lastName",
- * "phone",
- * "title",
- * "organization",
- * "city",
- * "region",
- * "country",
- * "zip",
- * "image",
- * "timezone"])
- * -------------------------------------------
- * The above call will map the fields other than the
- * exclusion list from the given keys to the destination payload
- *
+ * @param {*} msgType
  */
-
-function extractCustomFields(message, destination, keys, exclusionFields) {
-  keys.map((key) => {
-    const messageContext = get(message, key);
-    if (messageContext) {
-      const objKeys = [];
-      Object.keys(messageContext).map((k) => {
-        if (exclusionFields.indexOf(k) < 0) {
-          objKeys.push(k);
-        }
-      });
-      objKeys.map((k) => {
-        if (!(typeof messageContext[k] === "undefined")) {
-          if (destination) {
-            destination[k] = get(messageContext, k);
-          } else {
-            destination = {
-              k: get(messageContext, k),
-            };
-          }
-        }
-      });
-    }
+function checkReservedKeywords(message, msgType) {
+  // properties, traits, contextualTraits are either undefined or object
+  const objArr = [message.properties, message.traits, message.context.traits];
+  objArr.forEach((obj) => {
+    checkForReservedKeywords(obj, msgType);
   });
-  return destination;
 }
-/**
- *
- * @param {*} message
- *
- * Use get-value to retrieve defined traits from message traits
- */
-function getDefinedTraits(message) {
-  const traitsValue = {
-    userId:
-      get(message, "userId") ||
-      get(message, "context.traits.userId") ||
-      get(message, "anonymousId"),
-    email:
-      get(message, "context.traits.email") ||
-      get(message, "context.traits.Email") ||
-      get(message, "context.traits.E-mail"),
-    phone:
-      get(message, "context.traits.phone") ||
-      get(message, "context.traits.Phone"),
-    firstName:
-      get(message, "context.traits.firstName") ||
-      get(message, "context.traits.firstname") ||
-      get(message, "context.traits.first_name"),
-    lastName:
-      get(message, "context.traits.lastName") ||
-      get(message, "context.traits.lastname") ||
-      get(message, "context.traits.last_name"),
-    name:
-      get(message, "context.traits.name") ||
-      get(message, "context.traits.Name"),
-    city:
-      get(message, "context.traits.city") ||
-      get(message, "context.traits.City"),
-    country:
-      get(message, "context.traits.country") ||
-      get(message, "context.traits.Country"),
-  };
-
-  if (
-    !get(traitsValue, "name") &&
-    get(traitsValue, "firstName") &&
-    get(traitsValue, "lastName")
-  ) {
-    traitsValue.name = `${get(traitsValue, "firstName")} ${get(
-      traitsValue,
-      "lastName"
-    )}`;
-  }
-  return traitsValue;
-}
-
-/**
- * To check if a variable is storing object or not
- */
-const isObject = (obj) => {
-  return type(obj) === "object";
-};
-
-/**
- * To check if a variable is storing array or not
- */
-const isArray = (obj) => {
-  return type(obj) === "array";
-};
-
-const isDefined = (x) => x !== undefined;
-const isNotNull = (x) => x !== null;
-const isDefinedAndNotNull = (x) => isDefined(x) && isNotNull(x);
-
-const getDataFromSource = (src, dest, properties) => {
-  const data = {};
-  if (isArray(src)) {
-    for (let index = 0; index < src.length; index += 1) {
-      if (properties[src[index]]) {
-        data[dest] = properties[src[index]];
-        if (data) {
-          // return only if the value is valid.
-          // else look for next possible source in precedence
-          return data;
-        }
-      }
-    }
-  } else if (typeof src === "string") {
-    if (properties[src]) {
-      data[dest] = properties[src];
-    }
-  }
-  return data;
-};
 
 const getConfigUrl = (writeKey) => {
   return CONFIG_URL.concat(CONFIG_URL.includes("?") ? "&" : "?").concat(
@@ -674,72 +393,26 @@ const getConfigUrl = (writeKey) => {
   );
 };
 
-const checkSDKUrl = () => {
+const getSDKUrlInfo = () => {
   const scripts = document.getElementsByTagName("script");
-  let rudderSDK = undefined;
-  let staging = false;
+  let sdkURL;
+  let isStaging = false;
   for (let i = 0; i < scripts.length; i += 1) {
     const curScriptSrc = removeTrailingSlashes(scripts[i].getAttribute("src"));
-    // only in case of staging SDK staging env will be set to true
-    if (
-      curScriptSrc &&
-      curScriptSrc.startsWith("http") &&
-      (curScriptSrc.endsWith("rudder-analytics.min.js") ||
-        curScriptSrc.endsWith("rudder-analytics-staging.min.js"))
-    ) {
-      rudderSDK = curScriptSrc;
-      if (curScriptSrc.endsWith("rudder-analytics-staging.min.js")) {
-        staging = true;
+    if (curScriptSrc) {
+      const urlMatches = curScriptSrc.match(
+        /^(https?:)?\/\/.*rudder-analytics(-staging)?(\.min)?\.js$/,
+      );
+      if (urlMatches) {
+        sdkURL = curScriptSrc;
+        isStaging = urlMatches[2] !== undefined;
+        break;
       }
-      break;
     }
   }
-  return { rudderSDK, staging };
+  return { sdkURL, isStaging };
 };
 
-/**
- * Using this function we can create a payload from a mapping object.
- * @param {*} object = {
-   traits:{
-     name: "abcd efgh",
-     address: {
-       city: "xyz"
-     }
-   }
-  }
- * @param {*} mapper = [
-  {
-    destKey: "userName",
-    sourceKeys: "traits.name",
-  },
-  {
-    destKey: "city",
-    sourceKeys: "traits.address.city",
-  },
-]
- * @returns {
-   userName : "abcd efgh",
-   city : "xyz"
- }
-
-*/
-const constructPayload = (object, mapper) => {
-  const payload = {};
-  if (object)
-    mapper.forEach((element) => {
-      if (!Array.isArray(element.sourceKeys)) {
-        payload[element.destKey] = get(object, element.sourceKeys);
-      } else {
-        for (let i = 0; i < element.sourceKeys.length; i += 1) {
-          if (get(object, element.sourceKeys[i])) {
-            payload[element.destKey] = get(object, element.sourceKeys[i]);
-            break;
-          }
-        }
-      }
-    });
-  return payload;
-};
 
 export {
   replacer,
@@ -747,30 +420,19 @@ export {
   getCurrentTimeFormatted,
   getJSONTrimmed,
   getJSON,
-  getRevenue,
   getDefaultPageProperties,
   getUserProvidedConfigUrl,
   findAllEnabledDestinations,
   transformToRudderNames,
   transformToServerNames,
   handleError,
-  rejectArr,
-  type,
-  flattenJsonPayload,
   checkReservedKeywords,
   getReferrer,
   getReferringDomain,
-  extractCustomFields,
-  getDefinedTraits,
-  isObject,
-  isArray,
-  isDefinedAndNotNull,
-  getDataFromSource,
   commonNames,
   removeTrailingSlashes,
-  constructPayload,
   getConfigUrl,
-  checkSDKUrl,
+  getSDKUrlInfo,
   notifyError,
   leaveBreadcrumb,
   get,
